Tidy up controller, use new state
This moves the controller out in to it's own file, and starts writing
ENI info to the allocator state rather than to the old ENI map path
